Add flexible audio handling for multiple TTS engines
