+++ app/src/main/java/com/example/kokoro82m/viewmodel/BookViewModel.kt
@@ -8,10 +8,13 @@ import ai.onnxruntime.OrtSession
 import com.example.kokoro82m.utils.AudioPlayer
 import com.example.kokoro82m.utils.AudioPlayerManager
 import com.example.kokoro82m.utils.InterpolationMode
+import com.example.kokoro82m.utils.KittenAudioPlayer
+import com.example.kokoro82m.utils.KokoroAudioPlayer
 import com.example.kokoro82m.utils.PhonemeConverter
 import com.example.kokoro82m.utils.PlayerState
 import com.example.kokoro82m.utils.PlaybackNotification
 import com.example.kokoro82m.utils.StyleLoader
+import com.example.kokoro82m.utils.TtsEngine
 import com.example.kokoro82m.utils.playBook
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.Job
@@ -35,13 +38,22 @@ class BookViewModel : ViewModel() {
 
     private var appContext: Context? = null
 
-    val audioPlayer = AudioPlayer(
-        scope = viewModelScope,
-        onStateChanged = { state ->
-            _playerState.value = state
-            appContext?.let { PlaybackNotification.update(it, state) }
+    private var _audioPlayer: AudioPlayer? = null
+    val audioPlayer: AudioPlayer
+        get() = _audioPlayer!!
+
+    private fun initializeAudioPlayer(engine: TtsEngine) {
+        _audioPlayer = when (engine) {
+            TtsEngine.KOKORO -> KokoroAudioPlayer(viewModelScope) { state ->
+                _playerState.value = state
+                appContext?.let { PlaybackNotification.update(it, state) }
+            }
+            TtsEngine.KITTEN -> KittenAudioPlayer(viewModelScope) { state ->
+                _playerState.value = state
+                appContext?.let { PlaybackNotification.update(it, state) }
+            }
         }
-    )
+    }
 
     private var playJob: Job? = null
 
@@ -75,11 +87,13 @@ class BookViewModel : ViewModel() {
         startLine: Int,
         bookUri: Uri?,
         context: Context,
+        engine: TtsEngine,
         usePregenerated: Boolean,
         onFinished: () -> Unit,
     ) {
         playJob?.cancel()
         appContext = context.applicationContext
+        initializeAudioPlayer(engine)
         AudioPlayerManager.player = audioPlayer
         PlaybackNotification.show(appContext!!, true)
         playJob = playBook(
@@ -105,7 +119,7 @@ class BookViewModel : ViewModel() {
     fun stopPlayback() {
         playJob?.cancel()
         playJob = null
-        audioPlayer.stop()
+        _audioPlayer?.stop()
         appContext?.let { PlaybackNotification.cancel(it) }
         AudioPlayerManager.player = null
     }
